requirements.txt
Django==1.6.5
django-disqus==0.4.3
Markdown==2.5.2
Pillow==2.6.1
Pygments==2.0.1
psycopg2==2.5.3
South==1.0.1
Unipath==1.0
argparse==1.2.1
basicauth==0.2
bpython==0.13.1
cloudinary==1.0.18
dj-database-url==0.3.0
dj-static==0.0.6
django-debug-toolbar==1.2.2
django-devserver==0.8.0
django-extensions==1.4.8
django-twitter-feed==0.2
gunicorn==19.1.1
httplib2==0.9
oauthlib==0.7.2
python-decouple==2.3
python-oauth2==0.7.0
python-twitter==2.0
pytz==2014.10
requests==2.4.3
requests-oauthlib==0.4.2
simplejson==3.6.5
six==1.8.0
sqlparse==0.1.14
static3==0.5.1
wsgiref==0.1.2
